Allow repeating messages with empty content

+++ bot/commands/speak.py
@@ -16,7 +16,10 @@ if TYPE_CHECKING:
     brief="general.speak",
     description="Make the bot say something. Unlike `say`, this will attempt to delete your command message first",
 )
-async def _say_cmd(ctx: commands.Context[Haruka], *, content: str) -> None:
+async def _say_cmd(ctx: commands.Context[Haruka], *, content: str = "") -> None:
+    if not content and not ctx.message.attachments:
+        raise commands.UserInputError
+
     files = [await attachment.to_file() for attachment in ctx.message.attachments]
     with contextlib.suppress(discord.HTTPException):
         await ctx.message.delete()
